Move FishPrint to Print dir
